# sqle/docs/swagger.yaml
basePath: /
definitions:
...
  v1.AuditPlanMetaV1:
    properties:
      audit_plan_params:
        items:
          $ref: '#/definitions/v1.AuditPlanParamResV1'
        type: array
      audit_plan_type:
        type: string
      audit_plan_type_desc:
        type: string
      instance_type:
        type: string
    type: object
...
  v1.AuditPlanParamResV1:
    properties:
      desc:
        type: string
      key:
        type: string
      type:
        enum:
        - string
        - int
        - bool
        type: string
      value:
        type: string
    type: object
...
  v1.DashboardProjectTipV1:
    properties:
      project_name:
        type: string
      unfinished_workflow_count:
        description: 只统计与当前用户相关的未完成工单
        type: integer
    type: object
  v1.DashboardResV1:
    properties:
      workflow_statistics:
        $ref: '#/definitions/v1.WorkflowStatisticsResV1'
        type: object
    type: object
...
  v1.ExplainClassicResult:
    properties:
      head:
        items:
          $ref: '#/definitions/v1.TableMetaItemHeadResV1'
        type: array
      rows:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
    type: object
...
  v1.GetDashboardProjectTipsResV1:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/v1.DashboardProjectTipV1'
        type: array
      message:
        example: ok
        type: string
    type: object
  v1.GetDashboardResV1:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/v1.DashboardResV1'
        type: object
      message:
        example: ok
        type: string
    type: object
...
  v1.TableColumns:
    properties:
      head:
        items:
          $ref: '#/definitions/v1.TableMetaItemHeadResV1'
        type: array
      rows:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
    type: object
  v1.TableIndexes:
    properties:
      head:
        items:
          $ref: '#/definitions/v1.TableMetaItemHeadResV1'
        type: array
      rows:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
    type: object
  v1.TableMeta:
    properties:
      columns:
        $ref: '#/definitions/v1.TableColumns'
        type: object
      create_table_sql:
        type: string
      indexes:
        $ref: '#/definitions/v1.TableIndexes'
        type: object
      message:
        type: string
      name:
        type: string
      schema:
        type: string
    type: object
  v1.TableMetaItemHeadResV1:
    properties:
      desc:
        type: string
      field_name:
        type: string
    type: object
...
  v1.WorkflowStatisticsResV1:
    properties:
      my_need_execute_workflow_number:
        type: integer
      my_need_review_workflow_number:
        type: integer
      my_on_process_workflow_number:
        type: integer
      my_rejected_workflow_number:
        type: integer
      need_me_to_execute_workflow_number:
        type: integer
      need_me_to_review_workflow_number:
        type: integer
    type: object
...
  v2.AffectRows:
    properties:
      count:
        type: integer
      err_message:
        type: string
    type: object
  v2.AuditPlanReportSQLResV2:
    properties:
      audit_plan_report_sql:
        example: select * from t1 where id = 1
        type: string
      audit_plan_report_sql_audit_result:
        items:
          $ref: '#/definitions/v2.AuditResult'
        type: array
      number:
        example: 1
        type: integer
    type: object
  v2.AuditPlanResV2:
    properties:
      audit_plan_cron:
        example: 0 */2 * * *
        type: string
      audit_plan_db_type:
        example: mysql
        type: string
      audit_plan_instance_database:
        example: app1
        type: string
      audit_plan_instance_name:
        example: test_mysql
        type: string
      audit_plan_meta:
        $ref: '#/definitions/v1.AuditPlanMetaV1'
        type: object
      audit_plan_name:
        example: audit_for_java_app1
        type: string
      audit_plan_token:
        example: it's a JWT Token for scanner
        type: string
      rule_template:
        $ref: '#/definitions/v2.RuleTemplateV2'
        type: object
    type: object
...
  v2.AuditResult:
    properties:
      level:
        example: warn
        type: string
      message:
        example: 避免使用不必要的内置函数md5()
        type: string
      rule_name:
        type: string
    type: object
...
  v2.GetAuditPlanAnalysisDataResV2:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/v2.TaskAnalysisDataV2'
        type: object
      message:
        example: ok
        type: string
    type: object
  v2.GetAuditPlanReportSQLsResV2:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/v2.AuditPlanReportSQLResV2'
        type: array
      message:
        example: ok
        type: string
      total_nums:
        type: integer
    type: object
  v2.GetAuditPlansResV2:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/v2.AuditPlanResV2'
        type: array
      message:
        example: ok
        type: string
      total_nums:
        type: integer
    type: object
...
  v2.PerformanceStatistics:
    properties:
      affect_rows:
        $ref: '#/definitions/v2.AffectRows'
        type: object
    type: object
...
  v2.RuleTemplateV2:
    properties:
      is_global_rule_template:
        type: boolean
      name:
        type: string
    type: object
  v2.SQLExplain:
    properties:
      classic_result:
        $ref: '#/definitions/v1.ExplainClassicResult'
        type: object
      err_message:
        type: string
      sql:
        type: string
    type: object
  v2.TableMetas:
    properties:
      err_message:
        type: string
      table_meta_items:
        items:
          $ref: '#/definitions/v1.TableMeta'
        type: array
    type: object
  v2.TaskAnalysisDataV2:
    properties:
      performance_statistics:
        $ref: '#/definitions/v2.PerformanceStatistics'
        type: object
      sql_explain:
        $ref: '#/definitions/v2.SQLExplain'
        type: object
      table_metas:
        $ref: '#/definitions/v2.TableMetas'
        type: object
    type: object
...
info:
  contact: {}
  description: This is a sample server for dev.
  license: {}
  title: Sqle API Docs
  version: "1.0"
paths:
...
  /v1/dashboard:
    get:
      description: get dashboard info
      operationId: getDashboardV1
      parameters:
      - description: filter project name
        in: query
        name: filter_project_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetDashboardResV1'
      security:
      - ApiKeyAuth: []
      summary: 获取 dashboard 信息
      tags:
      - dashboard
  /v1/dashboard/project_tips:
    get:
      description: get dashboard project tips
      operationId: getDashboardProjectTipsV1
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetDashboardProjectTipsResV1'
      security:
      - ApiKeyAuth: []
      summary: 获取dashboard项目提示列表
      tags:
      - dashboard
...
  /v2/projects/{project_name}/audit_plans:
    get:
      description: get audit plan info list
      operationId: getAuditPlansV2
      parameters:
      - description: project name
        in: path
        name: project_name
        required: true
        type: string
      - description: filter audit plan db type
        in: query
        name: filter_audit_plan_db_type
        type: string
      - description: fuzzy search audit plan name
        in: query
        name: fuzzy_search_audit_plan_name
        type: string
      - description: filter audit plan type
        in: query
        name: filter_audit_plan_type
        type: string
      - description: filter audit plan instance name
        in: query
        name: filter_audit_plan_instance_name
        type: string
      - description: page index
        in: query
        name: page_index
        required: true
        type: integer
      - description: size of per page
        in: query
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.GetAuditPlansResV2'
      security:
      - ApiKeyAuth: []
      summary: 获取扫描任务信息列表
      tags:
      - audit_plan
  /v2/projects/{project_name}/audit_plans/{audit_plan_name}/reports/{audit_plan_report_id}/sqls:
    get:
      description: get audit plan report SQLs
      operationId: getAuditPlanReportsSQLs
      parameters:
      - description: project name
        in: path
        name: project_name
        required: true
        type: string
      - description: audit plan name
        in: path
        name: audit_plan_name
        required: true
        type: string
      - description: audit plan report id
        in: path
        name: audit_plan_report_id
        required: true
        type: string
      - description: page index
        in: query
        name: page_index
        required: true
        type: integer
      - description: size of per page
        in: query
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.GetAuditPlanReportSQLsResV2'
      security:
      - ApiKeyAuth: []
      summary: 获取指定扫描任务的SQL扫描详情
      tags:
      - audit_plan
  /v2/projects/{project_name}/audit_plans/{audit_plan_name}/reports/{audit_plan_report_id}/sqls/{number}/analysis:
    get:
      description: get SQL explain and related table metadata for analysis
      operationId: getAuditPlantAnalysisDataV2
      parameters:
      - description: project name
        in: path
        name: project_name
        required: true
        type: string
      - description: audit plan name
        in: path
        name: audit_plan_name
        required: true
        type: string
      - description: audit plan report id
        in: path
        name: audit_plan_report_id
        required: true
        type: string
      - description: sql number
        in: path
        name: number
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.GetAuditPlanAnalysisDataResV2'
      security:
      - ApiKeyAuth: []
      summary: 获取task相关的SQL执行计划和表元数据
      tags:
      - audit_plan
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"